requirements.txt
requests>=2.31.0
pdfplumber>=0.10.0
anthropic>=0.40.0
pandas>=2.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
thefuzz>=0.20.0
python-Levenshtein>=0.23.0
